refactor: parameters for service grid
